Added empty function bodies to avoid linker error
This allows the linker to avoid an issue involving undefined
references to virtual tables. This was caused by the
Unicorn_Interface class not providing the implementation for
it's overriding functions. This only occurred because
Unicorn_Interface is the derived class from Emulator_Interface,
which is an abstract class with only pure virtual functions.
A function body for the Emulator_interface de-constructor was
also required.