Added Coefficients file from original ELMO in converted file format
The file format has been changed to JSON and labelled.
This allows for easier understanding of the meaning being the data.
This also lets any data to be altered as the meaning is given by the
labels, permitting adding new interaction terms as well as new
instructions and new instruction categories.