Updated the Coefficients file (coeffs.json)
The format of the file has been changed to help reduce the potential for
errors. The Instruction Categories and Coefficients have been merged to
ensure that they have the same headings. This will also allow storing
the constant value under the same heading and simplify the required
parsing code.

Also, the individual opcodes have now been added to the file.