Added a CMake generated Model factory class as well as a template model.
The model factory was added to assist with the initialisation of Models
so that the main logic does not need to be aware of which model(s) are
in use.

The CMake module added generates the Model_Factory class based on which
models are available and specified in the CMakeLists.txt. This allows
for selective compilation of models as well as making the process of
adding new models easier.

A template for what a model should look like has been added. This can be
used to assist in adding new models.