Implemented Assembly_Instruction.hpp
The details stored about an assembly instruction include the opcode and
the registers. The registers are stored in both human readable and
encoded binary form.

Boost::dynamic_bitset was used to store arbitrary length binary operands.
In future this dependency should be removed if possible.

All supporting functions where changed to reflect the data that is now
stored.

A dummy implementation of constructing an Assembly_Instruction was added
to Execution.hpp to test for compile errors. This will be removed later.