Re-designed src/Execution.hpp
Originally the Execution class stored a list of instruction. It now
captures the full processor pipeline. This required completely changing
the majority of the class.
Additionally this also allowed multiple data types to be stored within
the Execution to best suit the specific processor/simulator.
Expansive Doxygen comments were added to support this new design.