Reworked Execution class to add new functionality
Converted m_registers from a 2d vector to a vector of maps. This allows
the storing of register names.

Added ability to store the type of stage of a pipeline stage and
detect abnormal states (Stall, Flush).

Added ability to retrieve an Instruction.

Added code to record pipeline stages.

Added code to record registers.

Added functionality to detect register access within an instruction.

Added functionality to retrieve the value within a register.