Modified Coefficients class to be more generic.
The Coefficients are now retrieved in a much more general way and should
be able to resist changes in the file format easier.

Retrieving from the Coefficients has been centralised to one function to
reduce the scope for errors and inconsistencies.

get_intstruction_categrory() was made public as some models may need
this information.

get_intstruction_categrory() was also modified slightly in a
non-functional way.

This all opens up the class to more use cases.