Added Get_State_unsafe and Is_Normal_State_Unsafe()
This are identical to the excisting Get_State and Is_Normal_State
function but do not throw an error in the case of out of bounds access.
Instead they return as if the state was stalled.

This has been used to prevent out of bounds errors in the Power model.